+++ lib/notification_listener_service.dart
@@ -29,7 +29,8 @@ class NotificationListenerService {
   /// it will open the notification settings page and return `true` once the permission granted.
   static Future<bool> requestPermission() async {
     try {
-      return await methodeChannel.invokeMethod('requestPermission');
+      final result = await methodeChannel.invokeMethod('requestPermission');
+      return result ?? false;
     } on PlatformException catch (error) {
       log("$error");
       return Future.value(false);